cody: log if embeddings were used in feedback
This includes an additional field in the event payload when a user
clicks on the thumbs up/thumbs down button. The field is a boolean
"lastChatUsedEmbeddings" which is true if any of the context files came
from embeddings.

I am unsure if this is the best approach to add this. I augmented the
ContextFile interface to optionally specify a source. I then in the only
place we directly use embeddings I set the source field. Other sources
will not populate this field (yet). Note: things like the unified
fetcher do not set this, which we may want to follow up on.

Note that the current implementation of feedback will just send the full
latest transcript, no matter which message a user clicked on. As such we
only base embeddings use on the last message. This is a limitation in
the existing implementation.

Test Plan: I modified the logEvent function to just console.log. I then
clicked on a thumbs up button with and without context files and
manually inspected the payload for the "lastChatUsedEmbeddings" field.